Improve handling of invalid art URLs
Sometimes the streamer returns an art URL which looks valid (it's a
string starting with "http://", but will return a non-200 response. That
used to result in rendering an empty art container. It now renders the
same "no art" container as when the url is completely missing.

This resulted in the UI performing its own JavaScript HTTP request,
which means using Tauri's own fetch API to avoid security/CORS issues.